Moved eventclass and unitindex to subfolder events
decluttered utils by moving event-related files.
