+++ Catalog.lua
@@ -5,6 +5,6 @@ require('CeresStdLib.handle.objects')
 
 require('CeresStdLib.util.benchmark')
 require('CeresStdLib.util.shell')
-require('CeresStdLib.util.eventclass')
 require('CeresStdLib.util.world')
-require('CeresStdLib.util.unitevent')
+require('CeresStdLib.util.events.eventclass')
+require('CeresStdLib.util.events.unitdex')
