+++ handle/objects.lua
@@ -3,3 +3,4 @@ require('CeresStdLib.handle.rect')
 require('CeresStdLib.handle.timer')
 require('CeresStdLib.handle.unit')
 require('CeresStdLib.handle.unitgroup')
+require('CeresStdLib.handle.trigger')
